WIP: replacing EnumSet's with ElemType's
