+++ src/types.rs
@@ -1,8 +1,8 @@
 // use crate::restack::{RestackError};
-use crate::elem::{Elem, ElemSymbol};
+use crate::elem::{Elem, ElemType, ElemTypeError};
 
 // Stack, StackError, LineNo, 
-use crate::stack::{Location};
+// use crate::stack::{Location};
 
 use std::collections::BTreeMap;
 use std::cmp;
@@ -13,8 +13,8 @@ use std::fmt::{Display, Formatter};
 use std::marker::PhantomData;
 // use std::sync::Arc;
 
-use enumset::{EnumSet, enum_set};
-use serde::{Deserialize, Serialize};
+// use enumset::{EnumSet, enum_set};
+// use serde::{Deserialize, Serialize};
 // use serde_json::{Map, Number, Value};
 use thiserror::Error;
 
@@ -185,171 +185,6 @@ impl AnError for Empty {
 // - two categories of tests:
 //   + property tests for typing methods themselves
 //   + test that a function having a particular type -> it runs w/o type errors on such inputs
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub enum BaseElemType {
-    Any,
-    Concat,
-    Index,
-    Slice,
-    ElemSymbol(ElemSymbol),
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct ElemTypeInfo {
-    base_elem_type: BaseElemType,
-    location: Location,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct ElemType {
-    type_set: EnumSet<ElemSymbol>,
-    info: Vec<ElemTypeInfo>,
-}
-
-// Formatting:
-// ```
-// ElemType {
-//     type_set: {A, B, C},
-//     info: _,
-// }
-// ```
-//
-// Results in:
-// ```
-// {A, B, C}
-// ```
-impl Display for ElemType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f,
-               "{{{}}}",
-               self.type_set.iter()
-               .fold(String::new(),
-                     |memo, x| {
-                         let x_str: &'static str = From::from(x);
-                         if memo == "" {
-                            x_str.to_string()
-                         } else {
-                            memo + ", " + &x_str.to_string()
-                         }
-                    }
-               ))
-    }
-}
-
-#[cfg(test)]
-mod elem_type_display_tests {
-    use super::*;
-
-    #[test]
-    fn test_empty() {
-        let elem_type = ElemType {
-            type_set: EnumSet::empty(),
-            info: vec![],
-        };
-        assert_eq!("{}", format!("{}", elem_type));
-    }
-
-    #[test]
-    fn test_singleton() {
-        for elem_symbol in EnumSet::all().iter() {
-            let elem_type = ElemType {
-                type_set: EnumSet::only(elem_symbol),
-                info: vec![],
-            };
-            assert_eq!(format!("{{{}}}", Into::<&'static str>::into(elem_symbol)),
-                       format!("{}", elem_type));
-        }
-    }
-
-    #[test]
-    fn test_all() {
-        assert_eq!("{Unit, Bool, Number, Bytes, String, Array, Object, JSON}",
-                   format!("{}", ElemType::any(vec![])));
-    }
-}
-
-impl ElemSymbol {
-    pub fn elem_type(&self, locations: Vec<Location>) -> ElemType {
-        ElemType {
-            type_set: EnumSet::only(*self),
-            info: locations.iter()
-                .map(|&location|
-                     ElemTypeInfo {
-                         base_elem_type: BaseElemType::ElemSymbol(*self),
-                         location: location,
-                    }).collect(),
-        }
-    }
-}
-
-impl Elem {
-    pub fn elem_type(&self, locations: Vec<Location>) -> ElemType {
-        self.symbol().elem_type(locations)
-    }
-}
-
-impl ElemType {
-    fn from_locations(type_set: EnumSet<ElemSymbol>,
-                      base_elem_type: BaseElemType,
-                      locations: Vec<Location>) -> Self {
-        ElemType {
-            type_set: type_set,
-            info: locations.iter()
-                .map(|&location|
-                     ElemTypeInfo {
-                         base_elem_type: base_elem_type,
-                         location: location,
-                    }).collect(),
-        }
-    }
-
-    pub fn any(locations: Vec<Location>) -> Self {
-        Self::from_locations(
-            EnumSet::all(),
-            BaseElemType::Any,
-            locations)
-    }
-
-    pub fn concat_type(locations: Vec<Location>) -> Self {
-        Self::from_locations(
-            enum_set!(ElemSymbol::Bytes |
-                      ElemSymbol::String |
-                      ElemSymbol::Array |
-                      ElemSymbol::Object),
-            BaseElemType::Concat,
-            locations)
-    }
-
-    pub fn index_type(locations: Vec<Location>) -> Self {
-        Self::from_locations(
-            enum_set!(ElemSymbol::Array |
-                      ElemSymbol::Object),
-            BaseElemType::Index,
-            locations)
-    }
-
-    pub fn slice_type(locations: Vec<Location>) -> Self {
-        Self::concat_type(locations)
-    }
-
-    pub fn unify(&self, other: Self) -> Result<Self, ElemTypeError> {
-        let both = self.type_set.intersection(other.type_set);
-        if both.is_empty() {
-            Err(ElemTypeError::UnifyEmpty {
-                lhs: self.clone(),
-                rhs: other.clone(),
-            })
-        } else {
-            let mut both_info = self.info.clone();
-            both_info.append(&mut other.info.clone());
-            Ok(ElemType {
-                type_set: both,
-                info: both_info,
-            })
-        }
-    }
-}
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct TypeId {
@@ -691,6 +526,13 @@ impl Type {
             o_type: self.o_type.iter().chain(mut_offset_other.o_type.iter().skip(zip_len)).copied().collect(),
         })
     }
+
+    pub fn prepend_inputs(&mut self, num_copies: usize, elem_type: ElemType) -> () {
+        if 0 < num_copies {
+            let type_id = self.context.push(elem_type);
+            self.i_type = (1..num_copies).into_iter().map(|_| type_id).chain(self.i_type.into_iter()).collect()
+        }
+    }
 }
 
 // Formatting:
@@ -793,16 +635,6 @@ mod type_display_tests {
 // TODO: split up TypeError
 // TODO: add layers of detail to TypeIdMapGetUnknownTypeId
 
-
-#[derive(Debug, PartialEq, Error)]
-pub enum ElemTypeError {
-    #[error("ElemType::unify applied to non-intersecting types: lhs: {lhs:?}; rhs: {rhs:?}")]
-    UnifyEmpty {
-        lhs: ElemType,
-        rhs: ElemType,
-        // location: TyUnifyLocation,
-    },
-}
 
 #[derive(Debug, PartialEq, Error)]
 pub enum TypeIdMapError {
